requirements.txt
paramiko
python-dotenv
django
channels
daphne
django-cors-headers
pyjwt
